pyproject.toml
[tool.poetry]
name = "pff-factfinder"
version = "0.0.6"
description = ""
authors = ["Baiyue Cao <caobaiyue@live.com>", "Molly Graber <email@email.com>"]
packages = [
    { include = "factfinder/*" },
    { include = "factfinder/**/*" },
]

[tool.poetry.dependencies]
python = ">=3.9.1,<4.0.0"
pandas = "1.1.4"
census = "0.8.15"
cached-property = "1.5.2"
python-dotenv = "0.15.0"
typer = "0.3.2"
tqdm = "^4.56.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.9.3"
pdoc3 = "^0.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
